.github/actions/update-documentation/scripts/requirements.txt
dirsync==2.2.5
pyyaml==5.3.1
argparse==1.4.0